figured out how to schedule this job, will need to set the secrets in the repo though
